The API was very sluggish due to the cloning of Ops and Pipelines. This cloning was needed to avoid problems due to the sharing of Ops and Pipelines. Example 1:
    a = ls(...) | select(...)
    b = a | map(...)

Without cloning, the expression assigned to b actually modifies that of
a, since it is a's Pipeline being extended.

Example 2:

    inc = map(x: x + 1)
    run(gen(...) | inc | inc)

The pipeline passed to run includes the SAME inc twice. The state of the
inc op is modified by both executions.

So a lot of cloning was needed to avoid this. Each __or__ has to clone
BOTH arguments to avoid this problem.

This commit switches to a different strategy. __or__ creates a Node,
building a tree of all the ops. Nodes are immutable, so in the first
example, a can point to a Node(ls, select), and b is a different node,
pointing to a and to a map Op.

A pipeline is instantiated only on execution. The created pipeline is
used for that one execution only. The instantiation copies each op
encountered, taking care of the gen | inc | inc example.